flask_app/requirements.txt
Flask==2.1.1
Werkzeug==2.1.1
mysql-connector-python==8.0.27
psycopg2-binary==2.9.3
